+++ docs/_includes/projects/properties/declaretypescript/my-element.ts
@@ -13,7 +13,7 @@ export class MyElement extends LitElement {
       <p>prop1: ${this.prop1}</p>
       <p>prop2: ${this.prop2}</p>
       <p>prop3: ${this.prop3}</p>
-      <p>prop4[0]:</p>${this.prop4[0]}</p>
+      <p>prop4[0]: ${this.prop4[0]}</p>
       <p>prop5.subprop1: ${this.prop5.subprop1}</p>
     `;
   }